src/plugins/python_django/project_root/pyproject.toml
[project]
name = "000project_name000"
version = "0.1.0"
description = "Project {{ project_name }} has been generated with the CLI tool for Modular Structure Generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django>=5.2.3",
  "djangorestframework>=3.16.0",
  "djangorestframework-simplejwt>=5.5.0",
  "httpx>=0.28.1",
  "python-dotenv>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
  "pytest>=8.4.1",
  "pytest-django>=4.11.1",
  "pytest-cov>=6.2.1",
  "pytest-mock>=3.14.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.3.0",
]

[tool.ruff]
line-length = 120
indent-width = 2
select = [
  "E",   # Pycodestyle errors
  "F",   # Pyflakes
  "W",   # Pycodestyle warnings
  "I",   # Isort
  "C",   # Complexity
  "B",   # Bugbear
  "UP",  # Pyupgrade
]
ignore = ["E501"]
